fix: constraints: RV32 uart RX/TX pin assignments swapped
No uart was being detected on the disco for the MIV_RV32
designs using the bare metal examples.

The pin constraints were originally assigned based on a
name match with the FTDI chip signals documented in
the board schematic:
  - The fabric RX was assigned to Y21 (FTDI_UART_RXD)
  - The fabric TX was assigned to W21 (FTDI_UART_TXD)

Since the schematic names describe the signals from the FTDI chip's
perspective, this resulted in a conflict where the fabric TX drove
the FTDI TX line, and the fabric RX listened to the FTDI RX line
- resulting in no crossover.

The change has been tested on the MIV_RV32_CFG1
design with the bare metal systick-blinky example
and confirmed to work.

Sar-id: 143969

Signed-off-by: Brian Burke <Brian.Burke@microchip.com>